[6] High-fidelity regression tests with NASA checkcases (#71) #major
* #6 - Oblateness calcs now match Vallado

* #6 - Minor improvements and bug fixes. Added plotting utilities to profling script as a temporary sandbox. Cleaned up eoms

* #6 - Fixed an issue where DCM from_vectors method made the opposite expected DCM. All propagation methods now align. Fixed various minor issues. Clean up drag force calcs

* #6 - Fixed a sign error in KeplerianVoP da/dt equation

* #6 - Removed legendre approximation cause it was wrong. Exact implementation doesn't seem to have significant slowdowns though. Might have been previous use of mutables.

* #6 - Fixed a minor bug where oblateness order would increment incorrectly

* #6 - Cleaned up oblateness force a bit

* #6 - Updated OblatenessForceTest to use output values until source of discrepancy can be identified

* #6 - Added NASA 6DoF Checkcases

* #6 - First cut at database building for 6dof checkcases

* #6 - Checkcase database built. Added some utilities. Need to add ability to build out classes from database now

* #6: Fixed and reorganized checkcase models. Fixed some minor issues in dockerfile

* #6: Fixed minor error in json util test

* #6: Reorganized 6dof test cases to be more unified

* #6 - WIP: Added first cut at test with 6dof data

* # 6 - Fixed checkcases and database. Added some more united-math operations, filled out first cut at orbital comparison test

* #6 - Moved plotting utilities into their own file (from profiling). Added output plots to help examine 6dof cases. Fixed issues that didn't allow base units to be changed. Fixed issue with quantity test comparitors. Added gnuplot to dockerfile

* #6 - WIP: partially upgraded to mp-units 2.5.0. Added a couple utilities to Keplerian and a couple plots to plotting. Still working on figuring out propagation differences

* # 6 - WIP: getting all my stashes aligned and deleted

* #6 - Fixed an issue where helper classes were not compiled with tests

* #6 - Fixed minor error scaling issues in integrator. Found that nasa checkases only disagreed due to difference in mu

* #6 - Cleaned up orbital checkcase test to make it more modular. Removed a lot of ancilliary debugging code

* #6 - Significantly cleaned up comparison code and made some output summaries. All comparisons output plots now. Fixed some plotting bugs. Largely fixed oblatness calcs but there's still some small descrepencies that need to be looked into. Current output matches to ~ 2m

* #6 - Nbody calcs have sub-meter agreement with no changes! Wohoo!

* #6 - Removed all the checkcase csvs from tracking since the database seems stable

* #6 - Cleaned up atmospheric force calculations to be a bit more human-readable

* #6 - Updated AstrodynamicsSystem method names to be more clear

* #6 - Added remaining missing checkcase tests. Now all non-rotational orbital checkcases have a corresponding test

* #6 - Minor cleanup in OblatnessForce. Added pow function for unitless quantities

* #6 - Updated gravity model data

* #6 - Minor updates. Added coefficients for WSG84 model. Fixed J200-ICRF get_dcm

* #6 - Fixed a bunch of compilation issues. I have no idea how it was compiling before

* #6 - Fixed an issue where time was not properly updating the epoch at which eoms were calculated. Replaced oblateness Pnm calls with more efficient recursive function.

* #6 - Added geocentric coordinate class. Fixed ECI to ECEF conversion (it was backwards). Replaced Oblateness calc with recurrence relation format (it now matches)

* #6 - Cleaned up Oblateness calcs

* #6 - Added more plots to outputs

* #6 - Fixed various minor bugs. Cleaned up debugging options and such from testing

* #6 - Add todos

* #6 - Updated workflows to use new build/install paradigm

* #6 - Updated README

* #6 - Updated build/test workflow to manually activate the python env

* #6 - Updated build/test workflow to manually activate the python env

* #6 - Updated activate_env

* #6 - Updated build and test workflow

* #6 - Updated build and test workflow

* #6 - Fixed an issue where AoA was not calculated properly

* #6 - Updated build and test workflow

* #6 - Added a bunch of missing tests. Updated coverage to ignore test files and the test directory. Coverage now also ignores snapshot cause I don't want to test that and make Spacetrack mad

---------

+++ astrea/math/math/interpolation.cpp
@@ -62,7 +62,7 @@ std::vector<double> cubic_spline(const std::vector<double>& x, const std::vector
         v[ii + 1] = b[ii + 1] - b[ii] - h[ii] * v[ii] / u[ii];
     }
 
-    for (size_t ii = n - 2; ii >= 0; --ii) {
+    for (int ii = n - 2; ii >= 0; --ii) {
         z[ii + 1] = (v[ii + 1] - h[ii + 1] * z[ii + 2]) / u[ii + 1];
     }
